Fix bug
When the webpage loads next chapter from auto click, it no longer is able
to pause the interval task.

And pausing all existing interval tasks in content script is required
because it lets more than one interval tasks to be alive when the website
reloads, eg from loading the next chapter.